+++ inst/uno/interfaces/AMPL/uno_ampl.cpp
@@ -25,13 +25,7 @@ namespace uno {
       const AMPLModel model(model_name);
       Uno uno{};
       Result result = uno.solve(model, options);
-      if (result.optimization_status == OptimizationStatus::SUCCESS) {
-         // check result.solution_status
-      }
-      else {
-         // ...
-      }
-      if (options.get_bool("AMPL_write_solution_to_file")) {
+      if (options.get_bool("write_solution_to_file")) {
          model.write_solution_to_file(result);
       }
       // std::cout << "memory_allocation_amount = " << memory_allocation_amount << '\n';
@@ -65,45 +59,51 @@ int main(int argc, char* argv[]) {
       // the -AMPL flag indicates that the solution should be written to the AMPL solution file
       size_t offset = 2;
       if (argc > 2 && std::string(argv[2]) == "-AMPL") {
-         options.set_bool("AMPL_write_solution_to_file", true);
+         options.set_bool("write_solution_to_file", true);
          ++offset;
       }
       else {
-         options.set_bool("AMPL_write_solution_to_file", false);
+         options.set_bool("write_solution_to_file", false);
       }
-      // get the command line arguments (options start at index offset)
-      const auto command_line_options = Options::get_command_line_options(argc, argv, offset);
-      std::optional<std::string> optional_option_file{};
-      std::optional<std::string> optional_preset{};
-      for (const auto& [option_name, option_value]: command_line_options) {
-         if (option_name == "option_file") {
-            optional_option_file = option_value;
+      try {
+         // get the command line arguments (options start at index offset)
+         const auto command_line_options = Options::get_command_line_options(argc, argv, offset);
+         std::optional<std::string> optional_option_file{};
+         std::optional<std::string> optional_preset{};
+         for (const auto& [option_name, option_value]: command_line_options) {
+            if (option_name == "option_file") {
+               optional_option_file = option_value;
+            }
+            else if (option_name == "preset") {
+               optional_preset = option_value;
+            }
          }
-         else if (option_name == "preset") {
-            optional_preset = option_value;
+
+         // [optional] set options from an option file
+         if (optional_option_file.has_value()) {
+            Options::load_option_file(options, *optional_option_file);
          }
-      }
 
-      // [optional] set options from an option file
-      if (optional_option_file.has_value()) {
-         Options::load_option_file(options, *optional_option_file);
-      }
-
-      // [optional] set a preset
-      if (optional_preset.has_value()) {
-         Presets::set(options, *optional_preset);
-      }
-
-      // set the rest of the command line options (note: we add preset to the options for debugging purposes)
-      for (const auto& [option_name, option_value]: command_line_options) {
-         if (option_name != "option_file") {
-            options.set(option_name, option_value);
+         // [optional] set a preset
+         if (optional_preset.has_value()) {
+            Presets::set(options, *optional_preset);
          }
-      }
 
-      // solve the model
-      Logger::set_logger(options.get_string("logger"));
-      run_uno_ampl(model_name, options);
+         // set the rest of the command line options (note: we add preset to the options for debugging purposes)
+         for (const auto& [option_name, option_value]: command_line_options) {
+            if (option_name != "option_file") {
+               options.set(option_name, option_value);
+            }
+         }
+
+         // solve the model
+         Logger::set_logger(options.get_string("logger"));
+         run_uno_ampl(model_name, options);
+      }
+      catch (const std::exception& e) {
+         std::cout << "uno_ampl failed with the following error: " << e.what() << '\n';
+         return EXIT_FAILURE;
+      }
    }
    return EXIT_SUCCESS;
 }