+++ inst/uno/uno/linear_algebra/Vector.hpp
@@ -27,6 +27,29 @@ namespace uno {
       Vector(Vector<T>&& other) noexcept: vector(std::move(other.vector)) { }
       ~Vector() = default;
 
+      // specialized constructor for expressions
+      template<typename E, typename = void>
+      struct is_expression : std::false_type {};
+
+      template<typename E>
+      struct is_expression<E,
+          std::void_t<
+              decltype(std::declval<E>().size()),
+              decltype(std::declval<E>()[std::size_t{}])
+          >
+      > : std::true_type {};
+
+      template<typename Expression, typename = std::enable_if_t<
+            is_expression<std::decay_t<Expression>>::value &&
+            !std::is_same_v<std::decay_t<Expression>, Vector>
+         >
+      >
+      Vector(Expression&& expression): vector(expression.size()) {
+         for (size_t index: Range(expression.size())) {
+            (*this)[index] = expression[index];
+         }
+      }
+
       // copy assignment operator
       Vector<T>& operator=(const Vector<T>& other) {
          if (&other != this) {
@@ -93,9 +116,7 @@ namespace uno {
       }
 
       void fill(T value) {
-         for (size_t index: Range(this->size())) {
-            this->vector[index] = value;
-         }
+         this->view().fill(value);
       }
 
       void print(std::ostream& stream) const {
@@ -108,29 +129,36 @@ namespace uno {
 
       template <typename Expression>
       Vector<T>& operator=(Expression&& expression) {
-         view(this->data(), this->size()) = std::forward<Expression>(expression);
+         this->view() = std::forward<Expression>(expression);
          return *this;
       }
 
       template <typename Expression>
       Vector<T>& operator+=(Expression&& expression) {
-         view(this->data(), this->size()) += std::forward<Expression>(expression);
+         this->view() += std::forward<Expression>(expression);
          return *this;
       }
 
       template <typename Expression>
       Vector<T>& operator-=(Expression&& expression) {
-         view(this->data(), this->size()) -= std::forward<Expression>(expression);
+         this->view() -= std::forward<Expression>(expression);
          return *this;
       }
 
       void scale(T factor) {
-         view(this->data(), this->size()).scale(factor);
+         this->view().scale(factor);
       }
-
 
    protected:
       std::vector<T> vector;
+
+      VectorView<const T> view() const noexcept {
+         return uno::view(this->data(), this->size());
+      }
+
+      VectorView<T> view() noexcept {
+         return uno::view(this->data(), this->size());
+      }
    };
 
    // use && to allow temporaries (such as std::cout or logger DEBUG, WARNING, etc)